requirements.txt
elenium 
lxml
requests
tqdm
wget